fix: sync the last code

+++ src/Modules/Admin/admin.service.ts
@@ -46,10 +46,10 @@ class AdminService {
     res
       .status(200)
       .cookie("accesstoken", accesstoken, {
-        httpOnly: true, // Prevent client-side access
-        sameSite: "lax",
-        secure: false, // true in production
-        maxAge: 90 * 24 * 60 * 60 * 1000, // 90 days
+        httpOnly: true,
+        sameSite: process.env.NODE_ENV == "production" ? "none" : "lax",
+        secure: process.env.NODE_ENV == "production",
+        maxAge: 90 * 24 * 60 * 60 * 1000,
       })
       .json(SuccessResponse("تم تسجيل الدخول بنجاح", 200, { accesstoken }));
   };
@@ -62,8 +62,8 @@ class AdminService {
   logOut = async (req: Request, res: Response) => {
     res.clearCookie("accesstoken", {
       httpOnly: true, // Prevent client-side access
-      sameSite: "lax",
-      secure: true,
+      sameSite: process.env.NODE_ENV == "production" ? "none" : "lax",
+      secure: process.env.NODE_ENV == "production",
     });
 
     res.status(200).json(SuccessResponse("تم تسجيل الخروج بنجاح", 200));
@@ -191,7 +191,11 @@ class AdminService {
     if (!product) throw new BadRequestException("هذا المنتج غير موجود");
 
     // delete the image of the product
-    if (product.imagePath && product.imagePath != "NO path") {
+    if (
+      product.imagePath &&
+      product.imagePath != "NO path" &&
+      fs.existsSync(product.imagePath)
+    ) {
       fs.unlinkSync(product.imagePath as string);
     }
